mgr/smb: add find_in_store function to config_store.py
Add a new `find_in_store` function to the config store module. This
function can be called on any store to search for a particular object
using the FindParams, with an exact match of the subset of keys & values
in the dict. If the store is a FindingConfigStore this operation may be
able to be done more efficiently but we always can fall back to the
simple approach of reading in every object from the store & namespace.

Resolves: rhbz#2300005
Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit fd6d4a77c85218e6f37167e05c603a3a9d91a4a9)