mgr/call_home: refactor agent
This refactor allows for versatility and ease of maintenance,
as well as introduces logic to handle cooldown and stale requests of
log uploads.

Signed-off-by: Yaarit Hatuka <yhatuka@ibm.com>
(cherry picked from commit 0b9327bb5b4e75f26715966890b8239fa684985c)

When cherry-picking to the 9.0 branch a conflict occurred:

Conflicts:
	src/pybind/mgr/call_home_agent/module.py

We resolved it by keeping the changes in this patch and deleting the
original version.