Fix lifecycle transition of encrypted multipart objects.
Lifecycle transtion can copy objects to a different storage tier.
When this happens, since the object is repacked, the original
manifest is invalidated.  It is necessary to store a special
"parts_len" attribute to fix this.  There was code in PutObj
to handle this, but that was only used for multisite replication,
it is not used by the lifecycle transisiton code.  This fix
adds similar logic to the lifecycle transition code path to make the
same thing happen.

Resolves: rhbz#2262741
Resolves: rhbz#2269342

Signed-off-by: Marcus Watts <mwatts@redhat.com>
(cherry picked from commit 15ced6304247703e269e916cf0e15e8f5d3f323f)