mgr/smb: add TransactingConfigStore protocol
Add a new TransactingConfigStore protocol. This TransactingConfigStore is a
ConfigStore that provides a method for establishing a transaction.
What a transaction means can depend on the store but generally it is
intended to provide a way to use database transactions with the store
abstraction.

Resolves: rhbz#2300005
Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit 29ab71466be6651c40e5e8b421f4b096ef395d7a)