Fixes to first part copy object encryption fixes (consolidated).
Fix an interface problem connecting filter chains.  The "get" half does
does a dainty dance over buffers in place, but the "put" half whomps
over its data (also in place) destructively.  We must copy the buffers
so the dance and the whomp don't step on each others toes.

Set the compression attribute *after* copyobject.  It needs
to see the data before it can properly generate the attribute.

When copying an encrypted object, it is necessary to do
'copy data' on it in order to ensure that it is re-keyed.
For use with sse-s3 between buckets, this is particularly
important.

Resolves: rhbz#2343537

Fixes: https://tracker.ceph.com/issues/23264

Signed-off-by: Marcus Watts <mwatts@redhat.com>
(cherry picked from commit 1712d9ce8002b1bfdb4e4449261fe6c800a1e387)

Conflicts:
	src/rgw/rgw_crypt.h