python-common/service_spec: fix zonegroup_hostnames type
This was meant to be a list of strings. The only use for it
was passing it to the rgw module with a mon_command call
so mypy didn't notice. New work to use this field for
custom SANs entries into self-signed certs revealed the issue

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 7f2324fb5f27901c3c07abb7623e121608d28fbc)