rgw/storage class: don't store/report STANDARD storage class.
While 'STANDARD' is a valid storage class, it is not supposed
to ever be returned when fetching an object.  This change suppresses
storing 'STANDARD' as the attribute value, so that objects
explicitly created with 'STANDARD' will in fact be indistinguishable
from those where it was implicitly set.

Fixes: https://tracker.ceph.com/issues/67786
Resolves: rhbz#2343537

Signed-off-by: Marcus Watts <mwatts@redhat.com>
(cherry picked from commit 1846c68b724c16c7840a46264f69376e7a6b5659)