rgw/etags: etags should be quoted in CompleteMultipartUploadResult xml.
Etags in other parts of ceph are returned as quoted objects (""),
where they are returned as header fields in the response.
Comppleting a multipart upload returns the etag in the xml
output, and it turns out they should be quoted here as well.

There are clients in the field (such as R2) that depend on this behavior.

Fixes: https://tracker.ceph.com/issues/67688
Resolves: rhbz#2343537

Signed-off-by: Marcus Watts <mwatts@redhat.com>
(cherry picked from commit 2509edea69e04c98c932975fbf5b8bea147a9ec9)

Conflicts:
	src/rgw/rgw_rest_s3.cc