rgw/storage class.  Don't inherit storage class for copy object.
When an object is copied, it should only be depending on data
in the request to determine the storage class, and if it is
not specified, it should default to 'STANDARD'.  In radosgw,
this means that this is another attribute (similar to encryption)
that should not be merged from the source object.

Fixes: https://tracker.ceph.com/issues/67787

Signed-off-by: Marcus Watts <mwatts@redhat.com>
(cherry picked from commit 3d0d30a0935adad2582ca674b85b1c58cb44a5e9)

Resolves: rhbz#2300284