mgr/cephadm: adding HA support for mgmt-gateway and ouath2-proxy
adding HA support for mgmt-gateway and ouath2-proxy. In addition, logic
is add to prevent unnecessary daemons restart during mgr failover.
Previously, without the management gateway, some daemons, such as
Prometheus, had an explicit dependency on the manager because we
needed to point to the active manager. With the management gateway,
this explicit dependency is no longer necessary, as it automatically
handles routing requests to the active manager.

Signed-off-by: Redouane Kachach <rkachach@ibm.com>