cephadm: also stop tcmu-runner container in post-stop
When transitioning from the old way cephadm deployed iscsi to the
new way with tcmu-runner as a sidecar, we can end up in situations where
we see something like

Unit process 1762503 (tcmu-runner-ent) remains running after unit stopped.
Unit process 1762508 (tcmu-runner) remains running after unit stopped.

in the logs of the rbd-target-api container (which was the only systemd
unit deployed prior to making tcmu-runner a sidecar). When this happens,
the daemon will initially fail to start with

Control process exited, code=exited, status=219/CGROUP

which will cause cephadm to mark the upgrade as failed. In the background,
systemd will restart the iscsi service and it tends to come up fine, but
cephadm has already marked the upgrade as failed at this point. I believe
this is an issue with the new unit files setup not being equipped for
stopping the tcmu-runner container and is strictly an issue with
transitioning between the two setups. This patch adds explicit lines to
stop the tcmu-runner container if it is present to the unit.poststop
file to address this. Note this should only have an effect during the
upgrade. Once the upgrade of iscsi is complete, the existing sidecar
setup will allow systemd to handle stopping the tcmu-runner service
as part of stopping the main systemd unit.

Signed-off-by: Adam King <adking@redhat.com>

Resolves: rhbz#2338125
(cherry picked from commit b46d53d780004ee6722847a59f6518e3b37f397d)