cephfs client: add zero-copy capabilit
For writes, the incoming iovec buffers are pushed into the bufferlist.

For reads, an incoming iovec is used to communicate the desired read
length, but then on read complete, an iovec would be built from the
bufferlist and passed back to the client.

Fixes: https://tracker.ceph.com/issues/69919
Signed-off-by: Igor Golikov <igolikov@ibm.com>
(cherry picked from commit ea2c58c7524acea0217466da9af5eba916da87ac)
Resolves: rhbz#2353897