mgr/dashboard: use existing pools for cephfs vol creation
We can use the newly introduced data and metadata params to create a vol
with those pools.

UI is being intelligent by filtering out the used pools and only uses
the pools that are labeled by cephfs and also not in use. To figure out
a pool is in use or not, we are fetching the pool stats and checking its
used_bytes.

Note: Using ec pools for data pool layout is something discouraged
according to offical doc: https://docs.ceph.com/en/latest/cephfs/createfs/#creating-a-file-system
We can force it but for now I have disabled it entirely in the dashboard
unless people say its okay to do it.

One more extra thing I am doing here is to add a note on deleting a
filesystem that the underlying pools and mds daemons will be removed.

Resolves: rhbz#2251887

Fixes: https://tracker.ceph.com/issues/70600
Signed-off-by: Nizamudeen A <nia@redhat.com>
(cherry picked from commit 8628b46c987296d269aabca2bdb9acc1e10a050f)