rgw: add force option to `radosgw-admin object rm ...`
The `radosgw-admin object rm ...` sub-command will give up if it
determines that there's an issue with the head object. This can make
it difficult for an admin to clean up a bucket index when there's a
damaged or missing head object.

When the user adds the "--yes-i-really-mean-it" command-line option,
it enables the "force mode". The bucket index entry(ies) will be
removed. If the object being removed is the current version in a
versioned bucket, the appropriate changes to the OLH will take place.

Resolves: rhbz#2313022

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>
(cherry picked from commit 7e1ac21069e14a4d6c4d4748971943b7c8b8137c)

  Conflicts -- minor reworking:
    src/rgw/driver/rados/rgw_bucket.cc
    src/rgw/driver/rados/rgw_rados.cc
    src/rgw/driver/rados/rgw_sal_rados.cc
    src/rgw/driver/rados/rgw_sal_rados.h
    src/rgw/rgw_sal.h
    src/rgw/rgw_sal_dbstore.cc
    src/rgw/rgw_sal_dbstore.h
    src/rgw/rgw_sal_filter.cc
    src/rgw/rgw_sal_filter.h

Resolves: rhbz#2313022
Signed-off-by: matt benjamin <mbenjamin@redhat.com>