cephadm: skip port in use check for rgw
To support openstack upgrades that go from RHCS 4 -> 5.
In their setup RGW is bound to a port on a specific ip.
Cephadm by default checks the port for a daemon on every
ip. In RHCS 6 onward, the plan is to add patches to allow
the RGW service to be configured such that cephadm will
properly check the port only on an ip within the networks
specified in the rgw spec file, but for 5 this workaround
patch will be used.

Signed-off-by: Adam King <adking@redhat.com>

Resolves: rhbz#2356354