cephadm: add _vendor with dataclases
The cephadm binary needs to run on  RHEL 8 which
uses python 3.6 that does not contain dataclasses which
are included in smb.py. We're bringing this in as a
single file to our source tree as the cephadm binary's
normal build mechanism uses either pip or RPMs, neither
of which can access without building on an older distro
that has python 3.6. The backport version of
dataclasses was taken from
https://github.com/ericvsmith/dataclasses/tree/master

Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit ee9dd883b6af4a4d0250e103a60f0add0916ac17)

Resolves: rhbz#2351292