mgr/cephadm: do not delete smb fs cephx keys
This change effectively disables fencing for the smb service because
the previous attempt to implement fencing would destroy the only
cephx key. Deleting this key would prevent any smb service part of
the logical cluster from talking to cephfs, even ones that were not
to be fenced.

The whole concept of fencing and ranks needs a bit of a rethink in
regards to smb. For now, we're just going to rely on ctdb and not
cephadm for smb's HA.

Fixes: 60300360cc500091e9dadf929d00bb72afad033c
Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit d72f1512bc6c97ceb0f7a8c35fe405ca658081df)

Resolves: rhbz#2353137