mgr/cephadm: split scheduling and deploying function in _apply_spec
Right now this commit on its own doesn't actually accomplish much, but it
is the beginning of splitting _apply_spec into a portion that will be
done serially per service (scheduling) and a portion that will be done
in parallel by host (deployment). As this patch itself stands, all
is still serial and by service and we're just passing the necessary
args between the functions. It is worth noting, some of the returns
of tuples of empty lists _apply_spec is doing here don't really
make sense, but I wanted the type checker to be okay with each
commit and also wanted to logically separate this commit and
the introduction of the deploy and removal queues

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit d12718bc30202dc58419025554e2b8e96b370545)

Resolves: rhbz#2364414

Conflicts:
	src/pybind/mgr/cephadm/serve.py