mgr/cephadm: do removal of daemons needing fencing with other daemons
In order to speed up the process of fencing daemons, this
patch has the daemons that need to be fenced be reported
and removed in parallel. Additionally, we now do the fencing
of that service's rank map after all the deployment/removal
is completed instead of before. Since the setting up of the rank
map is done by the scheduler before deploying/removing daemons
and fencing is primarily about cleaning up keyring for daemons
on offline hosts, I think the change in ordering shouldn't
cause any issues

Signed-off-by: Adam King <adking@redhat.com>

Conflicts:
	src/pybind/mgr/cephadm/serve.py

Resolves: rhbz#2372349